requirements.txt
# app
fastapi==0.73.0
uvicorn==0.17.4
pydantic==1.9
email-validator==1.1.3
python-multipart==0.0.5
requests==2.27.1
sqlalchemy-json==0.5.0
python-dotenv==0.20.0
fastapi-mail==0.4.1

#auth
passlib==1.7.4
bcrypt==3.2.0
PyJWT==2.4.0

# db
databases[postgresql]==0.3.1
SQLAlchemy==1.4.27
pg8000==1.24.0